# Conflicts:
#	.gitignore
#	docker/compose.yaml
#	src/main/resources/static/asset-manifest.json
#	src/main/resources/static/index.html

+++ src/main/java/com/binn328/ym_player/Model/Chapter.java
@@ -6,6 +6,6 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class Chapter {
-    private long timestamp;
+    private int timestamp;
     private String title;
 }
